controller di user completo

+++ src/main/java/EPICODE/EPIC_BNB/services/UsersService.java
@@ -9,6 +9,7 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 
+import EPICODE.EPIC_BNB.entities.Role;
 import EPICODE.EPIC_BNB.entities.User;
 import EPICODE.EPIC_BNB.entities.payload.UserCreatePayload;
 import EPICODE.EPIC_BNB.exceptions.BadRequestException;
@@ -25,6 +26,7 @@ public class UsersService {
 			throw new BadRequestException("Email " + user.getEmail() + " already in use!");
 		});
 		User newUser = new User(u.getName(), u.getSurname(), u.getUsername(), u.getEmail(), u.getPassword());
+		newUser.setRole(Role.USER);
 		return usersRepo.save(newUser);
 	}
 
@@ -33,13 +35,14 @@ public class UsersService {
 			size = 10;
 		if (size > 100)
 			size = 100;
+
 		Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
 
 		return usersRepo.findAll(pageable);
 	}
 
 	public User findById(UUID id) throws NotFoundException {
-		return usersRepo.findById(id).orElseThrow(() -> new NotFoundException("Utete con Id:" + id + "non trovato!!"));
+		return usersRepo.findById(id).orElseThrow(() -> new NotFoundException("Utete con Id:" + id + " non trovato!!"));
 	}
 
 	public User findByEmail(String email) throws NotFoundException {
@@ -52,20 +55,40 @@ public class UsersService {
 				.orElseThrow(() -> new NotFoundException("Utete:" + username + "non trovato!!"));
 	}
 
-	public User findByIdAndUpdate(UUID id, User u) throws NotFoundException {
-		User found = this.findById(id);
+	public User findByUsernameAndUpdate(String username, UserCreatePayload u) {
+		User user = usersRepo.findByUsername(username)
+				.orElseThrow(() -> new NotFoundException("Utete:" + username + "non trovato!!"));
+		user.setName(u.getName());
+		user.setSurname(u.getSurname());
+		user.setUsername(u.getUsername());
+		user.setEmail(u.getEmail());
 
-		found.setId(id);
-		found.setName(u.getName());
-		found.setSurname(u.getSurname());
-		found.setEmail(u.getEmail());
-
-		return usersRepo.save(found);
+		return usersRepo.save(user);
 	}
 
-	public void findByIdAndDelete(UUID id) throws NotFoundException {
+	public User findByIdAndUpdate(UUID id, UserCreatePayload u) {
 		User found = this.findById(id);
-		usersRepo.delete(found);
+		if (found == null) {
+			throw new NotFoundException("Nessun utente trovato con id: " + id);
+		} else {
+			found.setId(id);
+			found.setName(u.getName());
+			found.setSurname(u.getSurname());
+			found.setUsername(u.getUsername());
+			found.setEmail(u.getEmail());
+			usersRepo.save(found);
+			return found;
+		}
+
+	}
+
+	public void findByIdAndDelete(UUID id) {
+		User found = this.findById(id);
+		if (found == null) {
+			throw new NotFoundException("Nessun utente trovato con id: " + id);
+		} else
+			usersRepo.delete(found);
+
 	}
 
 }
